requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core
azure-search-documents==11.6.0b1
openai
python-dotenv
pydantic
fastapi
azure-cognitiveservices-speech
python-multipart
azure-storage-blob
pydub
num2words
azure-ai-documentintelligence
azure-ai-textanalytics
azure-data-tables
bs4==0.0.2
lxml
azure-ai-vision-imageanalysis==1.0.0b2
validators
user_agents
passlib==1.7.4
pyjwt
bcrypt==4.2.1
